+++ packages/gov-ui-kit/.storybook/preview.tsx
@@ -13,7 +13,7 @@ const preview: Preview = {
         options: {
             storySort: {
                 method: 'alphabetical',
-                order: ['Docs', ['Open Design System', 'Installation'], 'Design Tokens', 'Components'],
+                order: ['Docs', ['Documentation', 'Installation', 'Changelog'], 'Design Tokens', 'Components'],
             },
         },
         backgrounds: {