+++ packages/gov-ui-kit/CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Implement `Tag` component
 
+### Changed
+
+-   Update library documentation
+-   Output `build.css` file to be able to use ODS library without TailwindCSS
+
 ## [1.0.5] - 2023-11-20
 
 ### Changed